Merged master and fully separated header and impl files

+++ models/correlospinmatrix_detector.cpp
@@ -29,6 +29,7 @@
 // Includes from libnestutil:
 #include "compose.hpp"
 #include "dict_util.h"
+#include "genericmodel_impl.h"
 #include "logging.h"
 
 // Includes from nestkernel:
@@ -38,6 +39,7 @@
 #include "arraydatum.h"
 #include "dict.h"
 #include "dictutils.h"
+#include "nest_timeconverter.h"
 
 void
 nest::register_correlospinmatrix_detector( const std::string& name )
@@ -390,7 +392,7 @@ nest::correlospinmatrix_detector::handle( SpikeEvent& e )
       // yet every impulse in the queue that is further in the past than
       // this minimum - tau_max cannot contribute to the count covariance
       long t_min_on = t_i_on;
-      for ( int n = 0; n < P_.N_channels_; n++ )
+      for ( size_t n = 0; n < P_.N_channels_; n++ )
       {
         if ( S_.curr_state_[ n ] )
         {
